favorites support

+++ HackIllinois/DataSources/HIEventDataSource.swift
@@ -25,45 +25,62 @@ final class HIEventDataSource {
 
         HIEventService.getAllLocations()
         .onCompletion { result in
-            switch result {
-            case .success(let containedLocations):
-
+            if case let .success(containedLocations) = result {
+                print("GET::LOCATIONS::SUCCESS")
                 HIEventService.getAllEvents()
                 .onCompletion { result in
                     if case let .success(containedEvents) = result {
-                        DispatchQueue.main.sync {
-                            do {
-                                let ctx = CoreDataController.shared.persistentContainer.viewContext
-                                try? ctx.fetch(locationsFetchRequest).forEach {
-                                    ctx.delete($0)
-                                }
-                                try? ctx.fetch(eventsFetchRequest).forEach {
-                                    ctx.delete($0)
-                                }
-                                var locations = [Location]()
-                                containedLocations.data.forEach { location in
-                                    locations.append( Location(context: ctx, location: location) )
-                                }
+                        print("GET::EVENTS::SUCCESS")
+                        HIEventService.getAllFavorites()
+                        .onCompletion { result in
+                            if case let .success(containedFavorites) = result {
+                                print("GET::FAVORITES::SUCCESS")
+
+                                var updatedEvents = [HIAPIEvent]()
                                 containedEvents.data.forEach { event in
-                                    let eventLocationIds = event.locations.map { Int16($0.locationId) }
-                                    let eventLocations = locations.filter { eventLocationIds.contains($0.id) }
-                                    _ = Event(context: ctx, event: event, locations: NSSet(array: eventLocations))
+                                    var event = event
+                                    event.favorite = containedFavorites.data.map { $0.eventId }.contains(event.id)
+                                    updatedEvents.append(event)
                                 }
-                                try ctx.save()
-                            } catch { }
+
+                                DispatchQueue.main.sync {
+                                    do {
+                                        let ctx = CoreDataController.shared.persistentContainer.viewContext
+                                        try? ctx.fetch(locationsFetchRequest).forEach {
+                                            ctx.delete($0)
+                                        }
+                                        try? ctx.fetch(eventsFetchRequest).forEach {
+                                            ctx.delete($0)
+                                        }
+                                        var locations = [Location]()
+                                        containedLocations.data.forEach { location in
+                                            locations.append( Location(context: ctx, location: location) )
+                                        }
+                                        var events = [Event]()
+                                        updatedEvents.forEach { event in
+                                            let eventLocationIds = event.locations.map { Int16($0.locationId) }
+                                            let eventLocations = locations.filter { eventLocationIds.contains($0.id) }
+                                            events.append( Event(context: ctx, event: event, locations: NSSet(array: eventLocations)) )
+                                        }
+                                        HILocalNotificationController.shared.scheduleNotifications(for: events)
+
+                                        try ctx.save()
+                                    } catch { }
+                                }
+                            }
+                            completion?()
+                            isRefreshing = false
                         }
+                        .authorization(HIApplicationStateController.shared.user)
+                        .perform()
+                    } else {
+                        completion?()
+                        isRefreshing = false
                     }
-                    completion?()
-                    isRefreshing = false
                 }
                 .authorization(HIApplicationStateController.shared.user)
                 .perform()
-
-            case .cancellation:
-                completion?()
-                isRefreshing = false
-            case .failure(let error):
-                print(error)
+            } else {
                 completion?()
                 isRefreshing = false
             }
